Confirm UI works
