+++ Assets/Scripts/Search_Controller.cs
@@ -42,7 +42,7 @@ public class Search_Controller : MonoBehaviour
         var clickedNode = searchNodes[index];
         if (main_controller.Start_Node == null)
         {
-            _setStart(clickedNode);
+            SetStart(clickedNode);
             
         } else {
             confirm_Controller.Set_Confirm_State(clickedNode);
@@ -71,7 +71,7 @@ public class Search_Controller : MonoBehaviour
         InputUpdate();
     }
 
-    public void _setStart(Node_controller startNode)
+    public void SetStart(Node_controller startNode)
     {
         main_controller.Start_Node = startNode;
         var textBox = button_results[0].GetComponentInChildren<TMP_Text>();
